Scale elements in the canvas
This allows for scaled resizing. It works pretty well, but there
are some issues with the resize_canvas function (the handler for
the tkinter resizing event). First it is called at the beginning
of the program, and I don't understand why. Also, the exit()
function does not appear to work.