Fix double redraw at the start of the game
lemin_tk was redrawing twice at the start of the game because
the canvas receives a "<Configure>" event the first time it is
updated. Then, there was the initial drawing of the screen.

Now the canvas is updated once manually after it is created at
the end of init_screen(), and the initial vda.update state is
U_NONE which makes more sense than U_REDRAW.