Compress maps when the screen is not big enough
The map is compressed simply by scaling down the coordinates, it is
set to be the lowest possible (meaning that the visu is gonna compress
as little as needed and will prioritize keeping proportions as intact
as possible).

The final rendition is not ideal. It works, and is better than
nothing, but this should be modified. Maybe make compression a parameter.

Let's just say that this is the minimum level and that one room per
square of the grid is the maximum level (or something nearing this since
it wont always be possible to have a rectangle where "w*h=nbr_rooms",
the goal will be to have "(w*h) - nbr_rooms" as little as possible.