Change step value when speeding up or down
The new value is a scaled version of the original one, so that it stays
at the same place on the link. The ants used to go over the target room
when speeding up because the value was not changed. For example, if the
framec was 100 before the speeding, and it was 10 after, then any step
value over 10 would go too far.