Use the original room structure directly
This is part of an effort to streamline the visualizer. Now instead of
duplicating the room structure in the vda (visual data structure) in the
visualizer, the original one is directly used.

The goal is ultimately to not duplicate anything in the vda, and
directly use the lmap (lemin map structure) when needed. Links will also
get the same treatment. A link structure will be created, and a shape
element will be added to it.

The ants and grid shapes will get their own structure where needed, but
not in the lmap, since the ants won't be necessary in the map editor and
that the grid is an option of the printing. Although, the grid is not
just that, it is also an really important part of the printing logic.