Split lemin_screen and lemin_player (unfinished)
This is not finished. lemin_screen needs to handle the window and
the printing of the map, nothing more. The structure will have to
be usable both with lemin_player and lemin_editor.

So everything ant related needs to be moved to lemin_player. Redraw
and update_screen functions need to be pimped with function from
lemin_player at initialisation of the screen (with function pointers,
the same way lemin_player specific event handlers are given to
lemin_screen).