Fix move_room error and compress_coordinates
The move_room function was returning "x, y" instead of "s_x, s_y" which
infortunate, since this is the coordinates of the room. Also, compress
coordinates was compressing too far on some map. Which I never noticed
until now... weird... oh maybe this due to resolution of the screen which
is different here than it is at 42 (this is important because it is taken
into account when setting grid_min and when compressing of course).