Put new map button above edit map
Also, retrieve the name of the file with the return of the editor. This
does not work because of some weird things going on with the callback of
the buttons. It does not wait for the editor to be finished (which is a
good thing actually, since it allows running multiple instances the
lemin_editor or the lemin_player). And when it is done, the callback
does not "finish executing" the function. When closing the window, which
should lead the editor to return the wanted value, it is still not
updated in the menu. Only at the end is the function "finished". This is
a problem.