Test python's fork
This should be possible to use for sub windows in the menu. Multiprocessing
should also be looked at. There are some ways to connect between processes.

Ultimately the goal is to be able to launch as many games and/or editors at
the same time wihtout them affecting each other, and retrieve data at the end
in the parent process (this would typically be a string for the name of the
file created if any, to add it to the comboboxes in the menu).